Add Coding Guidelines and Helpful Resources
With this commit, the README now contains both a coding guidelines and
helpful resources section.  These sections are not complete and should
be populated with more information later, but the guidelines and links
contained within this commit should be a good starting point for
creating a common point of reference.