Add skeleton for feedback functionality
With this commit, the feedback module is mostly setup such that the
pseudocode is in place and the architecture is updated.  A new directory
has been created to house feedback strings.  This keeps the clutter of
the src/ directory down.  Access it like any other Python module
(example in Chef.py).