Implement LoadingScreen class
With this commit, a loading screen class is created with a run()
function dedicated to running the loading screens before the main game
loop.  Older commits that leveraged feedback have had their code removed
to refactor the functionality to LoadingScreen.  Temporary loading
screen memes have been added to allow the program to run.  They should
be replaced with the actual loading screens in a later commit.