Render Hasher in Front of Chef
With this commit, the register renders in front of the chef so it looks
like he walks behind the counter.  This creates incontinuity when the
chef looks like he's walking in front of the counter, but sticking to
the gameplay route does not reveal this. The starting coordinate for the
countertops have been shifted to the right to minimize this incontinuity
happening during normal gameplay.

The hasher has been refactored into its own rendering group so the
windows still register behind the chef.  Functionality appears to remain
the same otherwise.

fixes #108