Add end condition; Add misc enhancements
With this commit, a GameState enum is created to represent the different
stages of the game.  This variable is used as a reference for the main
run loop to decide whether or not to keep running.  A mock win screen
has been created as a placeholder for another commit to actually
implement it.

Other technical debt has been resolved, ranging from eliminating some
magic numbers to creating an app feedback messages file to extract magic
strings.

Technical debt created includes:
- Adding the win screen to app.py required importing new modules.
Perhaps refactor loading_screen.py to be a general module that allows
for the game to switch between panes conveniently.